refactor(vfs/syscall)：迁移statfs
