+++ kernel/src/syscall/mod.rs
@@ -346,12 +346,6 @@ impl Syscall {
                 return ret;
             }
 
-            SYS_STATFS => {
-                let path = args[0] as *const u8;
-                let statfs = args[1] as *mut PosixStatfs;
-                Self::statfs(path, statfs)
-            }
-
             SYS_FSTATFS => {
                 let fd = args[0] as i32;
                 let statfs = args[1] as *mut PosixStatfs;
